style: added new styles to user page
